test for login page
